anonymous-board/src/test/java/com/example/anonymousboard/support/token/JwtTokenProviderTest.java
package com.example.anonymousboard.support.token;

import static org.assertj.core.api.Assertions.assertThat;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.security.Keys;
import java.nio.charset.StandardCharsets;
import javax.crypto.SecretKey;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.boot.test.context.SpringBootTest;

@SpringBootTest
class JwtTokenProviderTest {

    @Autowired
    JwtTokenProvider jwtTokenProvider;

    @Value("${security.jwt.token.secret-key}")
    String secretKey;

    @DisplayName("JWT 토큰을 생성할 수 있다.")
    @Test
    void createAccessToken() {
        // given
        SecretKey key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
        Long id = 100000000000L;

        // when
        String accessToken = jwtTokenProvider.createAccessToken(id);
        Claims claim = Jwts.parserBuilder()
                .setSigningKey(key)
                .build()
                .parseClaimsJws(accessToken)
                .getBody();

        // then
        assertThat(claim.get("id")).isEqualTo(id);
    }
}